requirements.txt
django~=5.1
djangorestframework~=3.15
django-filter~=24.3
hypercorn~=0.17
psycopg2-binary~=2.9